move itemSelector to custom folder
