+++ DashAI/front/src/components/custom/ItemSelector.jsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import {
   Grid,
-  // DialogContentText,
   Typography,
   Paper,
   List,
